ai-built-agents/2025-12-26-daily-focus-enhancer/pyproject.toml
[tool.poetry]
name = "daily-focus-enhancer"
version = "0.1.0"
description = "An agent to enhance daily focus and productivity."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = ">=2.9.0"
langgraph = ">=0.2.0"
langchain-openai = ">=0.2.0"
python-dotenv = ">=1.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
